Fix /that bug/.
The GSA gets recognised as being a "sammy_ver1", with a screen resolution of 90x40. core.middleware decides that the maps given to this device should be 80x30. maps.draw.get_fitted_map insisted on a border of 20x35 to that markers didn't hang over the edge. This reduced the bounding box for markers to 60x(-5), which you can't fit any markers in. We entered an infinite loop on lines 205-206 as we reduced the zoom level in the hope that they would fit.

Now corrected to make the bounding box at least 64x64, regardless of the map size.