Robust handling of OSM tile fetching
If a tile can not be fetched, it is retried. If this fails, then a
fallback image is used (currently molly/osm/fallback/fallback.png).
The generated map is then marked as "faulty", and any future requests
for that map cause it to be regenerated, rather than expiring as per usual.