Add route planning based on timetables database
This is not ready for production and should be considered as early development code. There are known edge cases (journeys crossing midnight) not handled and performance is currently very slow for anything but the most trivial journeys.

Route planning currently considers entity groups (i.e., NaPTAN stop groups) as interchange points, however this is poor for multi-modal planning, therefore a more generous approach to interchanges needs to be undertaken.